feat: add ability to automatically decompress gzip responses returned from `fetch`
This patch add a new `fastly` options object to the RequestInit parameter on `fetch` and the `Request` constructor.
The new `fastly` options object has one field, `decompress`, if this field is falsey, then no automatic decompression is applied on the response. If this field is truthy, then automatic decompression is applied on the response.

This patch brings in support for automatically decompressing GZIP and not other compression formats.

I've added this under a field named `fastly` as that can be our extension point for other Fastly specific request/response features.

Previously we've added Fastly specific features as a root property on the RequestInit object, such as `backend`, `cacheOverride`, and `cacheKey`. In future patches we could move those to exist under `fastly` instead of the root.